refactor(cannon): move cannon utilities to Integration layer, update namespaces and imports, add Bannerlord.Cannons package reference, clean up test cases and project paths.
