+++ src/DellarteDellaGuerra.Infrastructure/MbObjects/IMBObjectProvider.cs
@@ -1,6 +1,6 @@
 using TaleWorlds.ObjectSystem;
 
-namespace DellarteDellaGuerra.Infrastructure;
+namespace DellarteDellaGuerra.Infrastructure.MbObjects;
 
 public interface IMBObjectProvider<T> where T : MBObjectBase
 {
