+++ src/DellarteDellaGuerra.Integration/DellarteDellaGuerra.Integration.csproj
@@ -19,6 +19,7 @@
         <PackageReference Include="IsExternalInit" Version="1.0.3" PrivateAssets="all" IncludeAssets="runtime; build; native; contentfiles; analyzers; buildtransitive"/>
         <PackageReference Include="NLog" Version="5.3.4"/>
 
+        <PackageReference Include="Bannerlord.Cannons" Version="0.3.0"/>
         <PackageReference Include="Harmony.DependencyInjection" Version="0.2.0"/>
 
         <PackageReference Include="Bannerlord.ReferenceAssemblies.Core" Version="$(GameVersion).*-*" PrivateAssets="All"/>
