+++ src/DellarteDellaGuerra.Integration/SiegeEngines/Util/ICannonAvailabilityProvider.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using TaleWorlds.CampaignSystem.Party;
 using TaleWorlds.Core;
 
-namespace DellarteDellaGuerra.Integration.SiegeEngines.Port
+namespace DellarteDellaGuerra.Integration.SiegeEngines.Util
 {
     public interface ICannonAvailabilityProvider
     {
