+++ src/DellarteDellaGuerra.Integration/SiegeEngines/Util/ICannonPrefabProvider.cs
@@ -1,6 +1,6 @@
 using TaleWorlds.Core;
 
-namespace DellarteDellaGuerra.Integration.SiegeEngines.Port
+namespace DellarteDellaGuerra.Integration.SiegeEngines.Util
 {
     public interface ICannonPrefabProvider
     {
