+++ src/DellarteDellaGuerra.Integration/SubModule.cs
@@ -11,17 +11,18 @@ using DellarteDellaGuerra.Domain.Common.Logging.Port;
 using DellarteDellaGuerra.Domain.DisplayCompilingShaders;
 using DellarteDellaGuerra.Domain.SiegeEngines;
 using DellarteDellaGuerra.Firearm;
+using DellarteDellaGuerra.Infrastructure.Configuration.Providers;
+using DellarteDellaGuerra.Infrastructure.Events;
+using DellarteDellaGuerra.Infrastructure.ExpandedTemplateApi.Logging;
+using DellarteDellaGuerra.Infrastructure.MbObjects;
+using DellarteDellaGuerra.Infrastructure.SiegeEngines.Port;
+using DellarteDellaGuerra.Infrastructure.Utils;
+using DellarteDellaGuerra.Integration.DI;
+using DellarteDellaGuerra.Integration.SiegeEngines;
 using DellarteDellaGuerra.Integration.SiegeEngines.Campaign;
 using DellarteDellaGuerra.Integration.SiegeEngines.Mission.Battle;
 using DellarteDellaGuerra.Integration.SiegeEngines.Mission.Siege.UI;
-using DellarteDellaGuerra.Infrastructure;
-using DellarteDellaGuerra.Infrastructure.Configuration.Providers;
-using DellarteDellaGuerra.Integration.DI;
-using DellarteDellaGuerra.Infrastructure.Events;
-using DellarteDellaGuerra.Infrastructure.ExpandedTemplateApi.Logging;
-using DellarteDellaGuerra.Infrastructure.SiegeEngines.Port;
-using DellarteDellaGuerra.Integration.SiegeEngines;
-using DellarteDellaGuerra.Infrastructure.Utils;
+using DellarteDellaGuerra.Integration.SiegeEngines.Util;
 using DellarteDellaGuerra.RemoveOrphanChildren.MissionBehaviours;
 using DellarteDellaGuerra.Tournament.Api;
 using DellarteDellaGuerra.Utils;
